refactor(webrtc): split rtc-peer-connection.ts method groups (978 → 634 + 4 modules)
Second slice of the `rtc-peer-connection.ts` split, sibling to PR #287
(which extracted the five JSEP entry points to
`rtc-peer-connection/sdp-negotiation.ts`). Picks up the four follow-up
method groups PR #287 explicitly listed as candidates:

  • `createDataChannel`
    → `packages/web/webrtc/src/rtc-peer-connection/data-channel.ts`
    (~134 LoC, including the `coerceUnsignedShort` Web-IDL helper
     moved alongside its only caller)

  • `addTransceiver`
    → `packages/web/webrtc/src/rtc-peer-connection/transceivers.ts`
    (~178 LoC, carrying both `request_pad_simple`-implicit-transceiver
     path A and `add-transceiver`-action-signal path B)

  • `addTrack`, `removeTrack`, `getSenders`, `getReceivers`,
    `getTransceivers`
    → `packages/web/webrtc/src/rtc-peer-connection/tracks.ts`
    (~121 LoC — the read-only accessors come along to keep the
     conceptual grouping intact)

  • `getStats`, `restartIce`, `setConfiguration`, `getConfiguration`
    → `packages/web/webrtc/src/rtc-peer-connection/stats-and-config.ts`
    (~132 LoC — they share the "live tuning without touching the
     media-pipeline construction path" theme)

11 methods totalling ~344 LoC moved out of the base verbatim. Same
`install*Methods(proto)` shape as PR #287 / WebGL2 (#273) /
canvas2d-core (#262) splits: a typed `<Group>Methods` interface
declaration-merged into `RTCPeerConnection` via `declare module
'../rtc-peer-connection.js' { ... }`, a `const <Group>Methods` object
literal with `ThisType<RTCPeerConnection>` carrying the verbatim method
bodies, plus an `install<Group>Methods(proto)` function that copies
them onto the prototype.

Base file `rtc-peer-connection.ts` 978 → 634 LoC. After the class
declaration it now calls all five install functions:

    installSdpNegotiationMethods(RTCPeerConnection.prototype);
    installDataChannelMethods(RTCPeerConnection.prototype);
    installTransceiverMethods(RTCPeerConnection.prototype);
    installTrackMethods(RTCPeerConnection.prototype);
    installStatsAndConfigMethods(RTCPeerConnection.prototype);

Each install function has both a bare side-effect import line and a
named import. tsc strips value-only named imports from emitted `.d.ts`,
so the side-effect imports are required to preserve the `declare module`
augmentations downstream — same pattern as PR #287.

Internal-field demotions extend PR #287's set (which had already
demoted `_pipeline`, `_webrtcbin`, `_iceRestartNeeded`, `_hasNegotiated`,
`_rejectIfClosed`, `_setStructureField`). This PR adds the following to
the package-internal (no `private` modifier, `_`-prefixed) convention so
the four new modules can reach them through `this: RTCPeerConnection`
typing:

    _conf, _closed, _dataChannels, _transceivers, _senders, _receivers,
    _applyIceServers, _applyIceTransportPolicy, _findNewGstTransceiver,
    _ensureTransports, _ensureSctpTransport, _createTransceiverWrapper,
    _handleNegotiationNeeded

The remaining state (`_bridge`, `_iceTransport`, `_dtlsTransport`,
`_sctpTransport`, `_handleIceCandidate`, `_handleNewTransceiver`,
`_handlePadAdded`, `_handleDataChannel`, `_dispatchStateChange`,
`_syncDtlsState`, `_syncIceState`, `_syncIceGatheringState`, `_on*`
handler slots) stays `private` — none of it is touched by any extracted
method.

With this PR the per-concern split of `RTCPeerConnection` is complete:
SDP negotiation, data-channel creation, transceiver factory, track
management, and stats/config now live in dedicated sibling modules; the
base file is back down to a manageable size (constructor, ICE-server /
policy config, state-property getters, close, transport-helper
utilities, signal handlers, on-event-attribute handlers).

Public surface preserved verbatim: every export from `@gjsify/webrtc`
works unchanged. Method signatures (`createDataChannel`,
`addTransceiver`, `addTrack`, `removeTrack`, `getSenders`,
`getReceivers`, `getTransceivers`, `getStats`, `restartIce`,
`setConfiguration`, `getConfiguration`) identical.

Test result: 101 GJS webrtc tests pass / 77 fail — identical to the
pre-refactor baseline (re-run via `git checkout main` to verify). All
77 failures are pre-existing, caused by `libnice-gstreamer1` not being
installed on this dev box; `ensureWebrtcbinAvailable()` throws
`NotSupportedError` before any extracted method runs. The extraction is
exercised at compile-time via `tsc --noEmit` (clean) and at runtime via
the five prototype-install side-effects preserved in the emitted
`.d.ts`.
